added support for symlink
